+++ ngcsimlib/__init__.py
@@ -1,5 +1,4 @@
 from ngcsimlib._src.component import Component as Component
-from ngcsimlib._src.process import Process
-
-import ngcsimlib.operations
-
+from ngcsimlib._src.process.methodProcess import MethodProcess
+from ngcsimlib._src.process.jointProcess import JointProcess
+from ngcsimlib._src.deprecators import deprecated, deprecate_args
